Territorios do saber -  Agrupar componentes curriculares
AB#11349 AB#11350 AB#11351 AB#11352 AB#11359 AB#11772 AB#11773 AB#11774

+++ src/SME.SGP.Api/Controllers/PlanoAnualController.cs
@@ -66,7 +66,7 @@ namespace SME.SGP.Api.Controllers
         [HttpGet("turmas/copia")]
         [ProducesResponseType(typeof(PlanoAnualCompletoDto), 200)]
         [ProducesResponseType(typeof(RetornoBaseDto), 500)]
-        public async Task<IActionResult> ObterTurmasParaCopia([FromQuery] int turmaId, [FromQuery] int componenteCurricular, [FromServices]IConsultasPlanoAnual consultasPlanoAnual)
+        public async Task<IActionResult> ObterTurmasParaCopia([FromQuery] int turmaId, [FromQuery] long componenteCurricular, [FromServices]IConsultasPlanoAnual consultasPlanoAnual)
         {
             return Ok(await consultasPlanoAnual.ObterTurmasParaCopia(turmaId, componenteCurricular));
         }
